Some minor refactoring to be a little more DRY :D
