+++ packages/Telephony/Actions/BatchTerminationCharacterInput.cs
@@ -5,10 +5,7 @@ using System.Runtime.CompilerServices;
 using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
-using AdaptiveExpressions.Properties;
 using Microsoft.Bot.Builder.Dialogs;
-using Microsoft.Bot.Builder.Dialogs.Adaptive;
-using Microsoft.Bot.Schema;
 using Newtonsoft.Json;
 
 namespace Microsoft.Bot.Components.Telephony.Actions
@@ -22,7 +19,7 @@ namespace Microsoft.Bot.Components.Telephony.Actions
         /// Class identifier.
         /// </summary>
         [JsonProperty("$kind")]
-        public new const string Kind = "Microsoft.Telephony.BatchTerminationCharacterInput";
+        public const string Kind = "Microsoft.Telephony.BatchTerminationCharacterInput";
 
         private const string _dtmfCharacterRegex = @"^[\d#\*]+$";
         private string _terminationCharacter;
@@ -40,15 +37,6 @@ namespace Microsoft.Bot.Components.Telephony.Actions
             // enable instances of this command as debug break point
             this.RegisterSourceLocation(sourceFilePath, sourceLineNumber);
         }
-
-        /// <summary>
-        /// Gets or sets intteruption policy. 
-        /// </summary>
-        /// <value>
-        /// Bool or expression which evalutes to bool.
-        /// </value>
-        [JsonProperty("allowInterruptions")]
-        public BoolExpression AllowInterruptions { get; set; } = true;
 
         /// <summary>
         /// Gets or sets the character that will be used to signal that the batch of input is complete.
@@ -79,29 +67,6 @@ namespace Microsoft.Bot.Components.Telephony.Actions
             {
                 return Task.FromResult(new DialogTurnResult(DialogTurnStatus.Waiting));
             }
-        }
-
-        /// <inheritdoc/>
-        protected override async Task<bool> OnPreBubbleEventAsync(DialogContext dc, DialogEvent e, CancellationToken cancellationToken)
-        {
-            if (e.Name == DialogEvents.ActivityReceived && dc.Context.Activity.Type == ActivityTypes.Message)
-            {
-                // Ask parent to perform recognition
-                await dc.Parent.EmitEventAsync(AdaptiveEvents.RecognizeUtterance, value: dc.Context.Activity, bubble: false, cancellationToken: cancellationToken).ConfigureAwait(false);
-
-                // Should we allow interruptions
-                var canInterrupt = true;
-                if (this.AllowInterruptions != null)
-                {
-                    var (allowInterruptions, error) = this.AllowInterruptions.TryGetValue(dc.State);
-                    canInterrupt = error == null && allowInterruptions;
-                }
-
-                // Stop bubbling if interruptions ar NOT allowed
-                return !canInterrupt;
-            }
-
-            return false;
         }
     }
 }
