+++ packages/Telephony/Common/RegexAggregatorInput.cs
@@ -9,7 +9,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using AdaptiveExpressions.Properties;
 using Microsoft.Bot.Builder.Dialogs;
-using Microsoft.Bot.Builder.Dialogs.Adaptive.Input;
+using Microsoft.Bot.Builder.Dialogs.Adaptive;
 using Microsoft.Bot.Schema;
 using Newtonsoft.Json;
 
@@ -20,11 +20,6 @@ namespace Microsoft.Bot.Components.Telephony.Actions
     /// </summary>
     public abstract class RegexAggregatorInput : Dialog
     {
-        /// <summary>
-        /// Class identifier.
-        /// </summary>
-        [JsonProperty("$kind")]
-        public const string Kind = "Microsoft.Telephony.RegexAggregatorInput";
         protected const string AggregationDialogMemory = "this.aggregation";
 
         /// <summary>
@@ -39,6 +34,15 @@ namespace Microsoft.Bot.Components.Telephony.Actions
             // enable instances of this command as debug break point
             this.RegisterSourceLocation(sourceFilePath, sourceLineNumber);
         }
+
+        /// <summary>
+        /// Gets or sets intteruption policy. 
+        /// </summary>
+        /// <value>
+        /// Bool or expression which evalutes to bool.
+        /// </value>
+        [JsonProperty("allowInterruptions")]
+        public BoolExpression AllowInterruptions { get; set; } = true;
 
         /// <summary>
         /// Gets or sets the regex pattern to use to decide when the dialog has aggregated the whole message.
@@ -109,6 +113,29 @@ namespace Microsoft.Bot.Components.Telephony.Actions
                 dc.State.SetValue(AggregationDialogMemory, existingAggregation);
                 return new DialogTurnResult(DialogTurnStatus.Waiting);
             }
+        }
+
+        /// <inheritdoc/>
+        protected override async Task<bool> OnPreBubbleEventAsync(DialogContext dc, DialogEvent e, CancellationToken cancellationToken)
+        {
+            if (e.Name == DialogEvents.ActivityReceived && dc.Context.Activity.Type == ActivityTypes.Message)
+            {
+                // Ask parent to perform recognition
+                await dc.Parent.EmitEventAsync(AdaptiveEvents.RecognizeUtterance, value: dc.Context.Activity, bubble: false, cancellationToken: cancellationToken).ConfigureAwait(false);
+
+                // Should we allow interruptions
+                var canInterrupt = true;
+                if (this.AllowInterruptions != null)
+                {
+                    var (allowInterruptions, error) = this.AllowInterruptions.TryGetValue(dc.State);
+                    canInterrupt = error == null && allowInterruptions;
+                }
+
+                // Stop bubbling if interruptions ar NOT allowed
+                return !canInterrupt;
+            }
+
+            return false;
         }
 
         private async Task<DialogTurnResult> PromptUserAsync(DialogContext dc, CancellationToken cancellationToken = default(CancellationToken))
